+++ frontend/src/components/search/AutoCompletedList.tsx
@@ -1,6 +1,6 @@
 import { Dispatch, SetStateAction, useEffect } from 'react';
 import styled from 'styled-components';
-import { MAX_TITLE_LENGTH } from '../constants/main';
+import { MAX_TITLE_LENGTH } from '../../constants/main';
 import { IAutoCompletedItem } from './Search';
 
 interface AutoCompletedListProps {
