+++ frontend/src/components/search/Search.tsx
@@ -2,15 +2,16 @@ import { ChangeEvent, KeyboardEvent, useCallback, useEffect, useMemo, useRef, us
 import { useQuery } from 'react-query';
 import { createSearchParams, useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
-import Api from '../api/api';
-import { PATH_SEARCH_LIST } from '../constants/path';
-import useDebounceValue from '../customHooks/useDebouncedValue';
-import MaginifyingGlassIcon from '../icons/MagnifyingGlassIcon';
-import { createDetailQuery } from '../utils/createQuery';
-import { getLocalStorage, setLocalStorage } from '../utils/localStorage';
+import Api from '../../api/api';
+import { PATH_SEARCH_LIST } from '../../constants/path';
+import useDebounceValue from '../../customHooks/useDebouncedValue';
+import MaginifyingGlassIcon from '../../icons/MagnifyingGlassIcon';
+import { createDetailQuery } from '../../utils/createQuery';
+import { getLocalStorage, setLocalStorage } from '../../utils/localStorage';
 import AutoCompletedList from './AutoCompletedList';
-import MoonLoader from './MoonLoader';
+import MoonLoader from '../MoonLoader';
 import RecentKeywordsList from './RecentKeywordsList';
+import IconButton from '../IconButton';
 
 enum DROPDOWN_TYPE {
   AUTO_COMPLETE = 'AUTO_COMPLETE',
@@ -194,9 +195,7 @@ const Search = ({ initialKeyword = '' }: SearchProps) => {
             onBlur={handleInputBlur}
             onKeyDown={handleInputKeyDown}
           />
-          <SearchButton type="button" onClick={() => handleSearchButtonClick(keyword)}>
-            <MaginifyingGlassIcon />
-          </SearchButton>
+          <IconButton icon={<MaginifyingGlassIcon />} onClick={() => handleSearchButtonClick(keyword)} />
         </SearchBar>
         <DropdownContainer>{renderDropdownContent(dropdownType)}</DropdownContainer>
       </SearchBox>
@@ -245,11 +244,6 @@ const SearchInput = styled.input`
   ::placeholder {
     color: ${({ theme }) => theme.COLOR.gray2};
   }
-`;
-
-const SearchButton = styled.button`
-  background-color: transparent;
-  cursor: pointer;
 `;
 
 const DropdownContainer = styled.div`
